use candidates file to build up contributions
