+++ .flake8
@@ -1,0 +1,8 @@
+[flake8]
+exclude =
+  venv,
+  **/migrations/*
+# So flake8 plays nicely with black
+# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html
+max-line-length = 120
+extend-ignore = E203
